Split views.rs into sub-modules
- src/views/mod.rs — main build() wiring
- src/views/gig_workspace.rs — notebook tabs, match results, auto-save
- src/views/contact_view.rs — contact panel builder and loader
- src/views/gig_sidebar.rs — sidebar row builder functions
- src/views/browser.rs — track/playlist/history list population
- src/views/dialogs.rs — settings and gig prep dialogs
- src/views/utils.rs — find_widget, fmt_time, rating_stars helpers
